+++ desktop/appcontainer/AboutDialog.js
@@ -17,11 +17,7 @@ import './AboutDialog.scss';
  * A dialog box showing basic metadata and version information about the Hoist application
  * and its plugins
  *
- * To add custom entries, specify in xhAboutMenuConfigs a list of objects in the form of
- * {"key": "configName", "label": "Display Name"}.
- *
- * Values for custom properties will be looked up via config key and fallback on AppModel.keyName
- * if no config exists.
+ * To customize the entries shown, override HoistAppModel.getAboutDialogItems()
  *
  * @private
  */